Issues #47 and #48 - Type migration of Encounter collections
A Contact's Encounters are now modeled as a SortedSet. Hibernate will now
rely on the Encounter's natural ordering when instantiating these
collections. The natural ordering relies on a known, application-wide
format for dates. This format is statically defined in the DateFormatConfig
class. Updated client code to supply dates in this format. Resolves #47
directly, and #48 seems to have been resolved by the type migration alone.