Issue #42 -- Front end support for entity updates and deletes
This commit provides client code for performing updates and deletes
on most of the core domain entities. The domain entities can now be
updated and deleted outright, and contacts can be removed from relations
with those entities as well.

Excluded from this commit are Encounter deletes. That functionality is
currently blocked by issue #49.