Issue #37 - Encounters should track follow ups
The Contact class now has a boolean property, needsFollowUp, that is set
by new Encounters that indicate a follow up is required. An Encounter
tracks whether the contact should be followed up by keeping a boolean
property, requiresFollowUp. If the new Encounter is also the most recent
Encounter, then the Contact's needsFollowUp flag will be updated;
otherwise, the Encounter will not affect this Contact field.