Issue #62 - Refactor Controller Exception Handling
This commit brings an implementation of a global @ControllerAdvice class
that handles custom exceptions thrown out of the controller layer.
These exceptions can, among other things, be mapped to specific Http status
codes. This is the groundwork for issuing clear responses to the user
when the app encounters basic constraint violations.