Issue #77 - Fixed test bugs
There was odd behavior in the CommitteeControllerTest that I decided to
move on from. Junit's After method was not being called after test methods
completed when built through maven, but was being called and tests were
passing when the test was run through our test suite runner. Moved
teardown logic to within the start of the test methods themselves as
workaround.