Issue #75 -- Include Group in the domain
Groups represent a collection of Aggregations. All of the existing domain
classes that contain collections of contacts - Committees, Organizations,
and Events -- now inherit from the Aggregation abstract class. This
polymorphic behavior allows us to create Groups composed of many different
types of constituents.

Introducing Groups to the app allows us to model, for example,
a simple hierarchy of an organization's committees.

Contacts are also allowed to be members of a Group and not necessarily
be a member of one of its constituent Aggregations.