Issue #76 -- Customizable, Discrete Sets of Demographic Category Options
A demographic attribute of Contact can now be set by referring to
the DemographicOptions of a DemographicCategory.

These sets of demographic options can be added to through the user inferface
which relies on its REST API. This allows the user to select from a defined
set of demographic options known to the system, and add to the set any
options that fit the Contact but aren't yet part of the system.