Travis/Heroku configuration - Issue#11
Created Procfile that, in coordination with travis-ci, will deploy
the application to a heroku instance after passing travis on the staging
branch.

Also included Spring Profiles to help in coordination with heroku
deployment. The application reads the spring.profiles.active environment
variable on startup to set the active profile. A 'dev' JPA config
class now exists to play well with the provisioned heroku DB. A 'production'
JPA config exists as before. The default active profile when no
environment variable is found is 'default'.