Closes #48 - Strip hard-coded implementation references
Several references to the SMWC implementation remained throughout the
project as hardcoded strings, organization logos, and credentials.

These references have been replaced with a default, implementation-agnostic
set of properties that are fetched by the front-end via a ConfigController.

This sets the stage for future work towards allowing greater flexibility in
startup configuration.