Closes #39 -- ApplicationStartup should draw user initialization properties from config file
On application startup, if and when users are created, their credentials are
now drawn from config.properties.

The basic user will now only be created on startup when the active Spring
profiles includes the dev profile. To take advantage of this, make sure
to set spring.profiles.active=dev in your environment/run configuration.

This pr also redefines what was previously called the dev spring
profile to what it actually is: a heroku spring profile.