Closes #54 -- API for Development domain
Basic CRUD for the development domain, excluding UserFileUploads, is now
provided by the REST layer of the app.

This includes creating and modifying Foundations, Grants, and
InteractionRecords; adding and modifying Donations from Contacts;
adding and removing particular Donations in isolation or relative to
an Organization or Event; and adding and modifying a Contact's
SustainerPeriods.

This API will need to be revisited at several points: when more nuanced
use cases are discovered that the API doesn't yet provide (likely very
soon); when we perform our 404 refactoring as per issue #58; and  when
we perform our validation refactoring as per issue #59.

This commit also includes BayardTestUtilities, a class that encompasses the
the logic behind MockMvc requests and offers a convenient way to validate
GET response bodies and validate POST/PUT/DELETE response statuses.

The new DtoTransformer class handles the simple mapping of domain entities
to and from their DTO representations.

The app now has additional dependency on Jackson's JSR310 module, which
handles the marshaling of Java's LocalDate and exposes some nice
configuration options for LocalDate JSON representation.