Closes #79 - Flag to enable/disable Development for client and server
Development features are now enabled by setting the
bayard.implementation.enableDevelopmentFeatures property to true.

If this property is set to anything other than true, the RestControllers
for the Development-related domain are not included in the application
context on startup, and requests to their mappings will return 404s.

The value of this property is communicated to the client along with the
other configuration options sent as a response from the ConfigController.
The client relies on a resolve function when the angular router tries to
navigate to one of the development-related URLs. If the root scope does
not yet contain the implementation config, the resolve functions fetches
the config from the server and stores it on the root scope. The resolve
function then allows or denies the route change to happen depending on
the value of the config object's developmentEnabled property.