Closes #74 -- Adds BudgetItem and InteractionRecordType
A donation's BudgetItem and an interaction's InteractionRecordType
are now proper types.

A user can create new BudgetItems and InteractionRecordTypes in the
new `Configuration` tab of the client app. The process of creating
EncounterTypes was moved under this tab.

Users can select one of a set of available options - ones created by
an administrator - when creating a new donation or a new interaction.