Closes #76 - Donor tab
There is now the simple beginnings of a Donor tab, intended to present
the user with an overview of its donors' activity. At this point, the tab
simply presents the user with the list of all contacts that are current
monthly sustainers. The page will support additional views of donor data
in time.