Closes #105 - Exposes created and lastModified timestamps for Contact
A Contact's created and lastModified times will now appear below
their new in their details page.

This commit also includes several revamps to various entity tables
throughout the applicaiton. Gone are any View Details buttons, replaced
by ng-clicks for table rows.

Also included is a revamping of each of our list pages. The tables in
these pages that expose all of a particular entity are now properly sized
in relation to the number of columns in the table. Entity creation
from these pages can now be done from a panel to the side of these tables
in a more aesthetic position.